requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
attrs==23.2.0
Authlib==1.3.0
bson==0.5.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
Faker==25.1.0
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi-login==1.10.0
fastapi-security==0.5.0
fonttools==4.51.0
future==1.0.0
geocoder==1.38.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
motor==3.4.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
trio==0.21.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
